Add --install-gsettings-schema CLI option
Add --install-gsettings-schema option that creates a
`com.github.epilys.rlr.Settings.gschema.xml` file at the given directory.
If directory is `-`, it will be printed to stdout instead.

The directory values should probably be one of:
["$HOME/.local/share/glib-2.0/schemas/", "/usr/share/glib-2.0/schemas/"]
in most systems.

As a reminder, the command

    glib-compile-schemas /path/to/glib-2.0/schemas/

Must be executed for changes to take effect.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>